+++ source/daemons/shadowd/shadowd.c
@@ -132,7 +132,8 @@ static int shadowd_is_old_master_enrolled(int sge_test_heartbeat, int sge_qmaste
       DRETURN(is_up_and_running);
    }
 
-   DPRINTF(("Try to send status information message to previous master host "SFQ" to port %ld\n", oldqmaster, sge_qmaster_port));
+   DPRINTF(("Try to send status information message to previous master host "SFQ" to port %ld\n",
+            oldqmaster, (long) sge_qmaster_port));
    ret = cl_commlib_get_endpoint_status(handle,oldqmaster ,(char*)prognames[QMASTER] , 1, &status);
    if (ret != CL_RETVAL_OK) {
       DPRINTF(("cl_commlib_get_endpoint_status() returned "SFQ"\n", cl_get_error_text(ret)));
